% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basemap_HH}
\alias{basemap_HH}
\title{Hamilton Harbour Base Map}
\format{
A ggmap raster object
}
\source{
Hamilton Harbour mapping data, Fish Ecology Science Lab (FESL)
}
\usage{
basemap_HH
}
\description{
A ggmap raster object containing a base map of Hamilton Harbour
for use as a background in spatial visualizations and plots.
}
\keyword{datasets}
