% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HH_shapefile_WGS84}
\alias{HH_shapefile_WGS84}
\title{Hamilton Harbour Shapefile (WGS84)}
\format{
An sf object with polygon geometry
}
\source{
Hamilton Harbour GIS data
}
\usage{
HH_shapefile_WGS84
}
\description{
An sf spatial object containing the shoreline boundary of
Hamilton Harbour in WGS84 coordinate reference system. Used for geographic
context in network plots.
}
\keyword{datasets}
