% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HHmapFull}
\alias{HHmapFull}
\title{Hamilton Harbour Full Map}
\format{
A ggmap object
}
\source{
Hamilton Harbour mapping data
}
\usage{
HHmapFull
}
\description{
A ggmap object containing a full base map of Hamilton Harbour
for visualization purposes.
}
\keyword{datasets}
