% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_use_template.R
\name{list_templates}
\alias{list_templates}
\title{List Available Templates}
\usage{
list_templates()
}
\value{
A character vector of available template names.
}
\description{
Shows all available analysis templates in the FESLtelemetry package.
}
\examples{
\dontrun{
list_templates()
}

}
