% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_analysis.R
\name{network_summary}
\alias{network_summary}
\title{Summarize Network Plot Data}
\usage{
network_summary(data, FishID, ReceiverID, lat, long, ...)
}
\arguments{
\item{data}{A dataframe containing acoustic telemetry detection data with fish IDs,
receiver IDs, and geographic coordinates.}

\item{FishID}{A vector or column from data describing fish identity (will be coerced to integer).}

\item{ReceiverID}{A vector or column from data describing acoustic receiver identity.}

\item{lat}{A numeric vector or column from data with latitude coordinates.}

\item{long}{A numeric vector or column from data with longitude coordinates.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list containing four elements:
\describe{
\item{receiver.locations}{Dataframe of unique receiver locations with detection frequencies}
\item{moves.matrix}{Matrix of pairwise movement counts between receivers}
\item{individual.moves}{Dataframe of individual fish movements (from, to, fish)}
\item{plot.data}{Dataframe formatted for network plotting with spatial coordinates}
}
}
\description{
Summarizes acoustic telemetry detection data to produce movement network
statistics and data structures for plotting. Tracks movements between receiver stations
and accounts for receivers that may have been relocated to multiple positions.
}
